Added a fix for getting unprefixed attribute values using JDK 1.6.0.18.  With the XMLStreamReader from that version, the calls
  getAttributeValue(null, "x")
  getAttributeValue("", "x")

both return "x1" for an element

  <elem ns1:x="x1" x="x2"/>

That is, it returns the first attributes whose name equals the given parameter
and ignores any prefixes.
The unprefixed attribute cannot be accessed. :-(

So now our XmlProcessor's getAttribute(String name) iterates the attributes
itself and picks the one that has no prefix and the given name.

fixes #221